requirements.txt
# FinPilot Dependencies
# Pinned for reproducibility - Sprint 3

# Web Framework
streamlit>=1.53.0,<2.0.0

# Data Processing
pandas>=2.3.0,<3.0.0
numpy>=2.4.0,<3.0.0
yfinance>=1.0.0,<2.0.0

# Visualization
plotly>=6.5.0,<7.0.0

# HTTP & Parsing
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.14.0,<5.0.0
lxml>=6.0.0,<7.0.0

# LLM Integration
google-generativeai>=0.8.0,<1.0.0
duckduckgo-search>=8.1.0,<9.0.0

# Core Infrastructure
pydantic>=2.12.0,<3.0.0
pydantic-settings>=2.12.0,<3.0.0
structlog>=25.0.0,<26.0.0

# Security - Sprint 1
pyjwt>=2.10.0,<3.0.0
bcrypt>=5.0.0,<6.0.0
cryptography>=46.0.0,<47.0.0

# Observability - Sprint 4
sentry-sdk>=2.31.0,<3.0.0
prometheus-client>=0.21.0,<1.0.0

# Export - Faz 2
openpyxl>=3.1.0,<4.0.0
reportlab>=4.2.0,<5.0.0

# Testing (dev)
pytest>=9.0.0
pytest-cov>=7.0.0
